requirements.txt
torch>=2.0
torchvision
numpy
pyyaml
tqdm
matplotlib
scipy
tensorboard
